Simple jobqueue using Operator SDK
Currently just a wrapper around Job.

Will be expanded to ensure that Jobs can pick work items off
the workqueue and process work items to completion.

As a job can be terninated for a variety of reasons including
- Priority
- Node Failure
- Resource Based termination
- Program error resulting in container termination

The goal is for Kubernetes to work in conjunction with the
job to ensure work items are handled as efficently as possible.

The controller with detect all failed pods and hand off workitems
assigned to the failed pod to newly created pods.

The controller does not care about the actual work. Just the workid
that is handed off to the pods.

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>